Revert "Moved Chat classes to com.pengrad.telegrambot.model.chat"
This reverts commit 8275d2ad93025b80ec0d42053caaef422c236e14.
