+++ library/src/main/java/com/pengrad/telegrambot/model/ChatMember.java
@@ -1,6 +1,4 @@
-package com.pengrad.telegrambot.model.chat;
-
-import com.pengrad.telegrambot.model.User;
+package com.pengrad.telegrambot.model;
 
 import java.io.Serializable;
 import java.util.Objects;
