+++ library/src/main/java/com/pengrad/telegrambot/model/chatboost/ChatBoost.java
@@ -1,6 +1,6 @@
-package com.pengrad.telegrambot.model.chat.boost;
+package com.pengrad.telegrambot.model.chatboost;
 
-import com.pengrad.telegrambot.model.chat.boost.source.ChatBoostSource;
+import com.pengrad.telegrambot.model.chatboost.source.ChatBoostSource;
 
 import java.util.Objects;
