+++ library/src/main/java/com/pengrad/telegrambot/response/GetChatResponse.java
@@ -1,6 +1,6 @@
 package com.pengrad.telegrambot.response;
 
-import com.pengrad.telegrambot.model.chat.Chat;
+import com.pengrad.telegrambot.model.Chat;
 
 /**
  * Stas Parshin
