+++ library/src/test/java/com/pengrad/telegrambot/ModelTest.java
@@ -1,7 +1,6 @@
 package com.pengrad.telegrambot;
 
 import com.pengrad.telegrambot.model.*;
-import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.request.InlineKeyboardButton;
 import com.pengrad.telegrambot.model.request.InlineKeyboardMarkup;
 import com.pengrad.telegrambot.passport.Credentials;
