+++ library/src/test/java/com/pengrad/telegrambot/TelegramBotTest.java
@@ -4,8 +4,6 @@ import com.pengrad.telegrambot.checks.*;
 import com.pengrad.telegrambot.impl.TelegramBotClient;
 import com.pengrad.telegrambot.model.*;
 import com.pengrad.telegrambot.model.botcommandscope.BotCommandScopeAllChatAdministrators;
-import com.pengrad.telegrambot.model.chat.*;
-import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.request.*;
 import com.pengrad.telegrambot.passport.*;
 import com.pengrad.telegrambot.request.*;
